requirements.txt
asgiref==3.9.1
Django==5.2.5
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
drf-nested-routers==0.94.2
mysqlclient==2.2.7
pillow==11.3.0
PyMySQL==1.1.1
python-decouple==3.8
sqlparse==0.5.3
